Pagina dipendenti fatta

+++ queries.js
@@ -21,9 +21,7 @@ function generateCodiceRitiro() {
 
 const getUsers = (request, response) => {
   pool.query('SELECT * FROM public.utenti ORDER BY id ASC', (error, results) => {
-    if (error) {
-      throw error
-    }
+    if (error) throw error
     response.status(200).json(results.rows)
   })
 }
@@ -32,62 +30,77 @@ const getUserById = (request, response) => {
   const id = parseInt(request.params.id)
 
   pool.query('SELECT * FROM public.utenti WHERE id = $1', [id], (error, results) => {
-    if (error) {
-      throw error
-    }
+    if (error) throw error
     response.status(200).json(results.rows)
   })
 }
 
 const creaCliente = (request, response) => {
-  const nome = request.body.nome;
-  const cognome = request.body.cognome;
-  const email = request.body.email;
-  const telefono = request.body.telefono;
-  const indirizzo = request.body.indirizzo;
-  const tipo = "CLIENTE";
-  const password = request.body.password;
-
   const salt = bcrypt.genSaltSync(10);
-  const hash = bcrypt.hashSync(password + "secret", salt);
-
-  console.log("\n\nsalt:", salt);
-  console.log("hash:", hash);
-
-  console.log("NUOVO UTENTE:\n", request.body);
+  const hash = bcrypt.hashSync(request.body.password + "secret", salt);
 
   pool.query('INSERT INTO public.utenti ( nome, cognome, email, password, salt, telefono, indirizzo, tipo) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)',
-    [nome, cognome, email, hash, salt, telefono, indirizzo, tipo], (error, results) => {
+    [request.body.nome, request.body.cognome, request.body.email, hash, salt, request.body.telefono, request.body.indirizzo, "CLIENTE"], (error, results) => {
       if (error) {
         throw error
       }
     })
+}
+
+const creaDipendente = (request, response) => {
+  const salt = bcrypt.genSaltSync(10);
+  const hash = bcrypt.hashSync(request.body.password + "secret", salt);
+  const decoded_token = jwt.decode(request.body.token_value);
+  var query, tipo;
+
+  switch (decoded_token.tipo) {
+    case "NEGOZIO":
+      tipo = "COMMERCIANTE";
+      query = 'INSERT INTO public.commercianti ( id, id_negozio ) VALUES ($1, $2)';
+      break;
+    case "MAGAZZINO":
+      tipo = "MAGAZZINIERE";
+      query = 'INSERT INTO public.magazzinieri ( id, id_magazzino ) VALUES ($1, $2)';
+      break;
+    case "DITTA_TRASPORTI":
+      tipo = "CORRIERE";
+      query = 'INSERT INTO public.corrieri ( id, id_ditta ) VALUES ($1, $2)';
+      break;
+  }
+
+  pool.query('INSERT INTO public.utenti ( nome, cognome, email, password, salt, telefono, indirizzo, tipo) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)',
+    [request.body.nome, request.body.cognome, request.body.email, hash, salt, request.body.telefono, request.body.indirizzo, tipo], (error, results) => {
+      if (error) throw error
+    })
+
+  findUserByEmail(request.body.email, (err, results) => {
+    if (err) return response.status(500).send('Server error!');
+
+    const dipendente = JSON.parse(JSON.stringify(results.rows));
+
+    if (dipendente.length == 1) {
+      pool.query(query, [dipendente[0].id, decoded_token.id], (error, results) => {
+        if (error) throw error
+      })
+      return response.status(200).send({ 'esito': "1" });
+    }
+    else return response.status(500).send("Server error!");
+  });
 }
 
 const creaAttivita = (request, response) => {
-  const ragione_sociale = request.body.ragione_sociale;
-  const tipo = request.body.tipo;
-  const email = request.body.email;
-  const telefono = request.body.telefono;
-  const indirizzo = request.body.indirizzo;
-  const password = request.body.password;
-
   const salt = bcrypt.genSaltSync(10);
-  const hash = bcrypt.hashSync(password + "secret", salt);
-
-  console.log("\n\nsalt:", salt);
-  console.log("hash:", hash);
-
-  console.log("NUOVA ATTIVITA:\n", request.body);
+  const hash = bcrypt.hashSync(request.body.password + "secret", salt);
 
   pool.query('INSERT INTO public.attivita ( ragione_sociale, tipo, email, password, salt, telefono, indirizzo ) VALUES ($1, $2, $3, $4, $5, $6, $7)',
-    [ragione_sociale, tipo, email, hash, salt, telefono, indirizzo], (error, results) => {
+    [request.body.ragione_sociale, request.body.tipo, request.body.email, hash, salt, request.body.telefono, request.body.indirizzo], (error, results) => {
       if (error) {
         throw error
       }
     })
 }
 
+//VECCHIO
 const updateUser = (request, response) => {
   const id = parseInt(request.params.id)
   const { nome, cognome } = request.body
@@ -104,15 +117,38 @@ const updateUser = (request, response) => {
   )
 }
 
-const deleteUser = (request, response) => {
-  const id = parseInt(request.params.id)
+const cercaDipendenteById = (id, decoded_token, cb) => {
+  var query;
+  switch (decoded_token.tipo) {
+    case 'NEGOZIO':
+      query = 'SELECT FROM public.commercianti WHERE id = $1 AND id_negozio = $2';
+      break;
+    case 'DITTA_TRASPORTI':
+      query = 'SELECT FROM public.corrieri WHERE id = $1 AND id_ditta = $2';
+      break;
+    case 'MAGAZZINO':
+      query = 'SELECT FROM public.magazzinieri WHERE id = $1 AND id_magazzino = $2';
+      break;
+  }
+  return pool.query(query, [id, decoded_token.id], (error, results) => {
+    cb(error, results)
+  });
+}
 
-  pool.query('DELETE FROM public.utenti WHERE id = $1', [id], (error, results) => {
-    if (error) {
-      throw error
-    }
-    response.status(200).send(`User deleted with ID: ${id}`)
-  })
+const eliminaDipendente = (request, response, decoded_token) => {
+  const id = parseInt(request.params.id);
+
+  cercaDipendenteById(id, decoded_token, (err, results) => {
+    if (err) return response.status(500).send('Server Error!');
+
+    const dipendente = JSON.parse(JSON.stringify(results.rows));
+    if (dipendente.length == 0) return response.status(404).send('Dipendente non trovato!');
+
+    pool.query('DELETE FROM public.utenti WHERE id = $1', [id], (error, results) => {
+      if (error) throw error
+      return response.status(200).send({ 'esito': "1" });
+    })
+  });
 }
 
 const findUserByEmail = (email, cb) => {
@@ -142,35 +178,34 @@ const creaOrdine = (request, response) => {
 }
 
 const getDipendenti = (token, cb) => {
-  //TODO: verificare il token!
-  const info = jwt.decode(token);
+  const decoded_token = jwt.decode(token);
+  var query;
 
-  switch (info.tipo) {
+  switch (decoded_token.tipo) {
     case 'NEGOZIO':
-      return pool.query(
-        'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.commercianti inner join public.utenti on public.commercianti.id=public.utenti.id where public.commercianti.id_negozio=$1', [info.id], (error, results) => {
-          cb(error, results)
-        });
+      query = 'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.commercianti inner join public.utenti on public.commercianti.id=public.utenti.id where public.commercianti.id_negozio=$1 ORDER BY cognome, nome ASC';
+      break;
     case 'DITTA_TRASPORTI':
-      return pool.query(
-        'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.corrieri inner join public.utenti on public.corrieri.id=public.utenti.id where public.corrieri.id_ditta=$1', [info.id], (error, results) => {
-          cb(error, results)
-        });
+      query = 'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.corrieri inner join public.utenti on public.corrieri.id=public.utenti.id where public.corrieri.id_ditta=$1 ORDER BY cognome, nome ASC';
+      break;
     case 'MAGAZZINO':
-      return pool.query(
-        'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.magazzinieri inner join public.utenti on public.magazzinieri.id=public.utenti.id where public.magazzinieri.id_magazzino=$1', [info.id], (error, results) => {
-          cb(error, results)
-        });
+      query = 'select public.utenti.id, nome, cognome, email, telefono, indirizzo from public.magazzinieri inner join public.utenti on public.magazzinieri.id=public.utenti.id where public.magazzinieri.id_magazzino=$1 ORDER BY cognome, nome ASC';
+      break;
   }
+
+  return pool.query(query, [decoded_token.id], (error, results) => {
+    cb(error, results)
+  });
 }
 
 module.exports = {
   getUsers,
   getUserById,
   creaCliente,
+  creaDipendente,
   creaAttivita,
   updateUser,
-  deleteUser,
+  eliminaDipendente,
   findUserByEmail,
   findAttivitaByEmail,
   creaOrdine,
